separate modules for exceptions, LR automaton
renaming of exceptions solves shadowing of builtin exceptions, but
creates potential incompatibility.
